requirements.txt
pygame
python-dotenv
numpy
